Check if source is a directory before symlink

+++ src/inc/commands.mjs
@@ -142,7 +142,7 @@ async function edit(slug) {
 	for (const source of getSourceDirectories()) {
 		const dirname = path.basename(source);
 		const symlink = `${ZXB_PATHS.sources}/${dirname}`
-		if (!fs.pathExistsSync(symlink) && fs.pathExistsSync(source)) {
+		if (!fs.pathExistsSync(symlink) && fs.pathExistsSync(source) && fs.isDirectorySync(source)) {
 			await $`ln -s ${source} ${symlink}`;
 		}
 	}
